introduce counter metrics and restructured go-template vars
prometheusMetricTemplates can now host multiple templates. templates get selected
based on their `stream` field and the stream label from the query publish command

the labels of the signalfx metric can now be access as .SignalFxLabels variable in
go-templates.

if a metric template does not specify a `name`, the signalfx metric name will be used as a default

examples have been revamped to show the use of counter metrics and metric template selection

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>