explict metric grouping configuration and group readiness (#17)
introduce configuration options for metric scraping based on labels to improve
blackbox-exporter compatibility mode. grouping options are now explicit and
include options to define when a group of metrics is ready for scraping, e.g.
number of metrics available within a group.

the async metric delivery mode of signalfx makes it necessary to handle situations
when metrics are yet missing (e.g. cold cache on process startup). the `groupReadyConditions` config
section provides options to declare the behaviour for such situations, failing a
scrape on the `/metric/$label_name` endpoint in situations when the conditions
are not satisfied. this will result in the default metric `up` on the scraper side
to highlight that metrics could not be aquired. depending on the situation, this
behaviour might be better than scraping incomplete or outdated metrics.

the following configuration enables filtered scraping based on the
specified label `label`. also, at least 2 metrics must be present after filtering
on the label. otherwise the scrape endpoint will fail with a descriptive error.

```yaml
grouping:
- label: label
  groupReadyConditions:
    minMetrics: 2
```

for filtering on the provided label, the url parameter `target` is used. this
provides compatibility with blackbox-exporter and enables the use of the `Probe`
CRD that comes with the Prometheus operator.

a valid scrape URL looks like this:

`:9091/metrics/label?target=value`

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>